update family model; update family CRUD operations; update family validation; improve family Interface and types

+++ src/logic/services/family.service.ts
@@ -1,24 +1,22 @@
-import { NotFoundException } from '../../utils/exceptions/index';
+import { ForbiddenException, NotFoundException } from '../../utils/exceptions/index';
 import { FamilyRepository } from '../../data/repositories/family.repository';
 import { FamilyResponseDto } from '../../logic/dtos/Family';
-import { FamilyEntity } from '../../data/entities/family.entity';
+import {
+  TCreateFamilyBody,
+  TFindFamiliesQuery,
+  TUpdateFamilyBody,
+} from '@/web/validators/family.validation';
 
 export class FamilyService {
-  static async getAll(): Promise<{ message: string; data: FamilyResponseDto[] }> {
-    const families = await FamilyRepository.find();
-    if (!families || families.length === 0) {
-      throw new NotFoundException('No families found');
-    }
-
-    return {
-      message: 'Families fetched',
-      data: FamilyResponseDto.fromMany(families),
-    };
-  }
-
-  static async create(createFamilyDto: any): Promise<{ message: string; data: any }> {
-    // const familyEntity = FamilyEntity.make(createFamilyDto);
-    const family = await FamilyRepository.create(createFamilyDto);
+  static async create(
+    ownerId: string,
+    familyData: TCreateFamilyBody,
+  ): Promise<{ message: string; data: any }> {
+    // !get app
+    // !get plan -> plan.maxSubs
+    // !create a subscription
+    // ! possible transactions {3}
+    const family = await FamilyRepository.create(familyData, ownerId);
 
     if (!family) {
       throw new NotFoundException('Failed to create family');
@@ -26,7 +24,22 @@ export class FamilyService {
 
     return {
       message: 'Family Created',
-      data: FamilyResponseDto.from(family.toObject()),
+      data: FamilyResponseDto.create(family.toObject()),
+    };
+  }
+
+  static async getAll(
+    filter: TFindFamiliesQuery,
+  ): Promise<{ message: string; data: FamilyResponseDto[] }> {
+    const families = await FamilyRepository.find(filter);
+    // *OpenApi spec: even if array is empty res.status is 200
+    if (!families || families.length === 0) {
+      throw new NotFoundException('No families found');
+    }
+
+    return {
+      message: 'Families fetched',
+      data: FamilyResponseDto.fromMany(families),
     };
   }
 
@@ -43,9 +56,9 @@ export class FamilyService {
   }
 
   static async getFamilyOwner(
-    owner: string,
+    reqUser: string,
   ): Promise<{ message: string; data: FamilyResponseDto[] }> {
-    const family = await FamilyRepository.findOwners({ owner });
+    const family = await FamilyRepository.find({ owner: reqUser });
     if (!family) {
       throw new NotFoundException('No family found');
     }
@@ -57,29 +70,34 @@ export class FamilyService {
   }
 
   static async getFamilySubscribers(
-    subscribers: string,
+    familyId: string,
   ): Promise<{ message: string; data: FamilyResponseDto[] }> {
-    const family = await FamilyRepository.getSubscribers({ subscribers });
+    const family = await FamilyRepository.findById(familyId);
     if (!family) {
       throw new NotFoundException('No family found');
     }
+    // !use subscriber service method to find subscribers with 'familyId'
 
     return {
       message: 'Family fetched',
-      data: FamilyResponseDto.fromMany(family),
+      data: FamilyResponseDto.fromMany([family]),
     };
   }
 
   static async update(
     familyId: string,
-    updateFamilyDto: any,
+    newData: TUpdateFamilyBody,
+    reqUser: string,
   ): Promise<{ message: string; data: FamilyResponseDto }> {
     const family = await FamilyRepository.findById(familyId);
     if (!family) {
       throw new NotFoundException('No family found');
     }
 
-    const updatedFamily = await FamilyRepository.update(familyId, updateFamilyDto);
+    if (!family.owner.equals(reqUser))
+      throw new ForbiddenException({ message: "you are not the family's owner" });
+
+    const updatedFamily = await FamilyRepository.update(familyId, newData);
 
     if (!updatedFamily) {
       throw new NotFoundException('No family found after update');
@@ -91,11 +109,18 @@ export class FamilyService {
     };
   }
 
-  static async delete(familyId: string): Promise<{ message: string; data?: FamilyResponseDto }> {
-    const family = await FamilyRepository.delete(familyId);
+  static async delete(
+    familyId: string,
+    reqUser: string,
+  ): Promise<{ message: string; data?: FamilyResponseDto }> {
+    const family = await FamilyRepository.findById(familyId);
     if (!family) {
       throw new NotFoundException('No family found');
     }
+    if (!family.owner.equals(reqUser))
+      throw new ForbiddenException({ message: "you are not the family's owner" });
+
+    await FamilyRepository.delete(familyId);
 
     return {
       message: 'Family deleted',
